moved unrelated scopes (Scope) to models
