+++ dal/backup/restore.py
@@ -13,7 +13,7 @@ import glob
 import shutil
 from datetime import datetime
 from zipfile import ZipFile, BadZipFile
-from dal.scopes import scopes
+from dal.models import scopes
 from movai_core_shared.logger import Log
