+++ dal/models/flow.py
@@ -10,7 +10,7 @@
 
 from types import SimpleNamespace
 from movai_core_shared.logger import Log
-from dal.scopes import scopes
+from .scopestree import scopes
 from dal.helpers import GFlow
 from dal.helpers.parsers import ParamParser
 from movai_core_shared.consts import (ROS1_NODELETSERVER)
