+++ dal/models/nodeinst.py
