+++ dal/plugins/persistence/filesystem/filesystem.py
@@ -20,8 +20,7 @@ from movai_core_shared.logger import Log
 from dal.plugins import Plugin
 from dal.data import (schemas, PersistencePlugin,
                       TreeNode, SchemaPropertyNode, Persistence)
-from dal.scopes import ScopeInstanceVersionNode, ScopesTree
-from dal.models import Model
+from dal.models import Model, ScopesTree, ScopeInstanceVersionNode
 from movai.remote import RemoteArchive
 from dal.backup import RestoreManager
