+++ dal/plugins/persistence/redis/redis.py
@@ -16,8 +16,7 @@ from redis.exceptions import ResponseError
 from dal.plugins import Plugin
 from dal.data import (Persistence, PersistencePlugin, SchemaPropertyNode,
                       SchemaNode, schemas, TreeNode)
-from dal.scopes import ScopesTree, ScopeInstanceVersionNode
-from dal.models import Model
+from dal.models import Model, ScopesTree, ScopeInstanceVersionNode
 from dal.movaidb import MovaiDB
